+++ solr1/local/92-init-solr.sh
@@ -9,6 +9,7 @@ set -e
 echo "Solr init start"
 
 SOLR_DATA_ROOT=${SOLR_DATA_ROOT:-'/opt/alfresco/alf_data/solr4'}
+DIR_ROOT=${DIR_ROOT:-'/opt/alfresco/alf_data'}
 
 JAVA_XMS=${JAVA_XMS:-'512M'}
 JAVA_XMX=${JAVA_XMX:-'2048M'}
@@ -174,10 +175,10 @@ echo "export JAVA_OPTS" >> $TOMCAT_CONFIG_FILE
 
 user="tomcat"
 # make sure backup folders exist and have the right permissions in case of mounts
-if [[ ! -d "${SOLR_DATA_ROOT}/solr4Backup" ]]
+if [[ ! -d "${DIR_ROOT}/solrBackup" ]]
 then
-    mkdir -p "${SOLR_DATA_ROOT}/solr4Backup/alfresco" "${SOLR_DATA_ROOT}/solr4Backup/alfresco"
-	chown -hR "$user":"$user" "${SOLR_DATA_ROOT}/solr4Backup"
+    mkdir -p "${DIR_ROOT}/solrBackup/alfresco" "${DIR_ROOT}/solrBackup/archive"
+	chown -hR "$user":"$user" "${DIR_ROOT}/solrBackup"
 fi
 # fix permissions for whole data folder in case of mounts
 if [[ $(stat -c %U /opt/alfresco/alf_data) != "$user" ]]